Incorporated meta key for depth alarms. //TODO get the config values to go automatically into self.json without having to do it explicitly in SignalKModelFactory
SignalKModelFactory.java
Generated meta entry for depth alarms from config data. This should be
handled in the Json-schema that generates the config form for the server.

ConfigConstants.java
Changed key for DEPTH_TRANSDUCER_OFFSET to conform to SignalK

JsonSerializer
Added the key for depth.belowTransducer.meta.zones. to the list of array keys.
Don't know if this is really necessary.