On startup, signalk-config.json and self.json are initialized with the default configuration values if they are not present. Restart Server button now enters the new config values into the model.
SignalKModelFactory.java
Added System.out.println debugging statements to the SignalK static block.
Log4J is initialized later.

Added loadConfig() to the load method so that when the system is initialized by 
signalk-config.json, the selected config parameters are injected into the 
SignalK model.

Broke up the loadConfig() method, moving some of the code to a new method
insertConfig that puts the config parameters into the SignalK model.

In the loadConfig method, if the signalk-config.json file does not exist, 
a call to Util.setDefaults() and a call insertConfig() ensures that the 
default parms get into self.json.